+++ meta/SM/beta_g3.m
@@ -9,5 +9,7 @@
      2 g3^6 (-63559 + 89896 Zeta[3]) + 
      27 (20 gt^4 \[Lambda] - 12 gt^2 \[Lambda]^2 + 
         gt^6 (141 + 8 Zeta[3])))
- ) /. Zeta[3] -> N[Zeta[3]]
-}
+ ),
+ (* from Baikov, Chetyrkin, Kühn [1606.08659] *)
+ -g3^11*(458001185 + 2831724*Pi^4 - 1420919880*Zeta[3] + 937800000*Zeta[5])/6480
+} /. Zeta[n_] :> N[Zeta[n]]