+++ model_specific/SM/standard_model.hpp
@@ -577,6 +577,7 @@ private:
    double calc_beta_g3_two_loop(const Beta_traces&) const;
    double calc_beta_g3_three_loop(const Beta_traces&) const;
    double calc_beta_g3_four_loop(const Beta_traces&) const;
+   double calc_beta_g3_five_loop(const Beta_traces&) const;
    double calc_beta_Lambdax_one_loop(const Beta_traces&) const;
    double calc_beta_Lambdax_two_loop(const Beta_traces&) const;
    double calc_beta_Lambdax_three_loop(const Beta_traces&) const;